+++ client/consensus/src/consensus_container.rs
@@ -30,8 +30,9 @@ use {
 };
 
 use {
+    crate::authorities,
     futures::{lock::Mutex, prelude::*},
-    nimbus_primitives::CompatibleDigestItem as NimbusCompatibleDigestItem,
+    nimbus_primitives::{CompatibleDigestItem as NimbusCompatibleDigestItem, NimbusPair},
     sc_client_api::{backend::AuxStore, BlockOf},
     sc_consensus::{BlockImport, BlockImportParams, ForkChoiceStrategy, StateAction},
     sc_consensus_aura::{find_pre_digest, CompatibilityMode},
@@ -39,12 +40,11 @@ use {
         BackoffAuthoringBlocksStrategy, SimpleSlotWorker, SlotInfo, SlotResult, StorageChanges,
     },
     sc_telemetry::{telemetry, TelemetryHandle, CONSENSUS_DEBUG, CONSENSUS_INFO, CONSENSUS_WARN},
-    sp_api::{Core, ProvideRuntimeApi},
+    sp_api::ProvideRuntimeApi,
     sp_application_crypto::{AppKey, AppPublic},
     sp_blockchain::HeaderBackend,
     sp_consensus::{
-        BlockOrigin, EnableProofRecording, Environment, Error as ConsensusError, ProofRecording,
-        Proposer, SyncOracle,
+        BlockOrigin, EnableProofRecording, Environment, ProofRecording, Proposer, SyncOracle,
     },
 };
 
@@ -566,115 +566,6 @@ where
             self.logging_target(),
         )
     }
-}
-
-fn authorities<P, B, C>(
-    client: &C,
-    parent_hash: B::Hash,
-    context_block_number: NumberFor<B>,
-    compatibility_mode: &CompatibilityMode<NumberFor<B>>,
-    keystore: SyncCryptoStorePtr,
-) -> Result<Vec<AuthorityId<P>>, ConsensusError>
-where
-    P: Pair + Send + Sync,
-    P::Public: AppPublic + Hash + Member + Encode + Decode,
-    P::Signature: TryFrom<Vec<u8>> + Hash + Member + Encode + Decode,
-    B: BlockT,
-    C: ProvideRuntimeApi<B>,
-    C::Api: TanssiAuthorityAssignmentApi<B, AuthorityId<P>>,
-    AuthorityId<P>: From<<NimbusPair as sp_application_crypto::Pair>::Public>,
-{
-    let runtime_api = client.runtime_api();
-
-    match compatibility_mode {
-        CompatibilityMode::None => {}
-        // Use `initialize_block` until we hit the block that should disable the mode.
-        CompatibilityMode::UseInitializeBlock { until } => {
-            if *until > context_block_number {
-                runtime_api
-                    .initialize_block(
-                        parent_hash,
-                        &B::Header::new(
-                            context_block_number,
-                            Default::default(),
-                            Default::default(),
-                            parent_hash,
-                            Default::default(),
-                        ),
-                    )
-                    .map_err(|_| sp_consensus::Error::InvalidAuthoritiesSet)?;
-            }
-        }
-    }
-    first_eligible_key::<B, C, P>(client.clone(), keystore.clone(), &parent_hash)
-        .ok_or(sp_consensus::Error::InvalidAuthoritiesSet)
-}
-
-use nimbus_primitives::{NimbusId, NimbusPair, NIMBUS_KEY_ID};
-/// Grab the first eligible nimbus key from the keystore
-/// If multiple keys are eligible this function still only returns one
-/// and makes no guarantees which one as that depends on the keystore's iterator behavior.
-/// This is the standard way of determining which key to author with.
-pub(crate) fn first_eligible_key<B: BlockT, C, P>(
-    client: &C,
-    keystore: SyncCryptoStorePtr,
-    parent_hash: &B::Hash,
-) -> Option<Vec<AuthorityId<P>>>
-where
-    C: ProvideRuntimeApi<B>,
-    C::Api: TanssiAuthorityAssignmentApi<B, AuthorityId<P>>,
-    P: Pair + Send + Sync,
-    P::Public: AppPublic + Hash + Member + Encode + Decode,
-    P::Signature: TryFrom<Vec<u8>> + Hash + Member + Encode + Decode,
-    AuthorityId<P>: From<<NimbusPair as sp_application_crypto::Pair>::Public>,
-{
-    // Get all the available keys
-    let available_keys = SyncCryptoStore::keys(&*keystore, NIMBUS_KEY_ID).ok()?;
-
-    // Print a more helpful message than "not eligible" when there are no keys at all.
-    if available_keys.is_empty() {
-        log::warn!(
-            target: LOG_TARGET,
-            "🔏 No Nimbus keys available. We will not be able to author."
-        );
-        return None;
-    }
-
-    let runtime_api = client.runtime_api();
-
-    // Iterate keys until we find an eligible one, or run out of candidates.
-    // If we are skipping prediction, then we author with the first key we find.
-    // prediction skipping only really makes sense when there is a single key in the keystore.
-    let maybe_authorities = available_keys.into_iter().find_map(|type_public_pair| {
-        // Have to convert to a typed NimbusId to pass to the runtime API. Maybe this is a clue
-        // That I should be passing Vec<u8> across the wasm boundary?
-        if let Ok(nimbus_id) = NimbusId::from_slice(&type_public_pair.1) {
-            // If we dont find any parachain that we are assigned to, return non
-
-            if let Ok(Some(para_id)) =
-                runtime_api.check_para_id_assignment(parent_hash.clone(), nimbus_id.clone().into())
-            {
-                log::error!("Para id found for assignment {:?}", para_id);
-                let authorities = runtime_api
-                    .para_id_authorities(parent_hash.clone(), para_id)
-                    .ok()?;
-                log::error!(
-                    "Authorities found for para {:?} are {:?}",
-                    para_id,
-                    authorities
-                );
-                authorities
-            } else {
-                log::error!("nO Para id found for assignment {:?}", nimbus_id);
-
-                None
-            }
-        } else {
-            None
-        }
-    });
-
-    maybe_authorities
 }
 
 /// Parameters of [`build_aura_worker`].
